First commit of the clonal algorithm code.
The algorithm isn't really developed yet.  I'm just getting familiar
with python again.  Have a loop creating 100 random 32bit integers in an
array, and then doing a little hamming distance stuff on the first two.
We'll probably use this type of code to implement the clonal selection
stuff like in "The Baldwin Effect in the Immune System" paper.