Basic functional interface
We could definitely wrap all of these in a class, but I don't know
if that's necessary for our limited scope.  Pretty easy to see what's
going on from the function names.  I'm not sure this code is capturing
the whole complexity of an AIS, but the basic GA structure is there.
Definitely have a ways to go, but I think we're off to a good start.